# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Storytail API",
        "description": "API documentation for Storytail",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/users/{userId}/books": {
            "get": {
                "tags": [
                    "User Books"
                ],
                "summary": "Lista de livros do utilizador",
                "description": "Retorna os livros lidos e favoritos de um utilizador específico",
                "operationId": "1390ea3bef4247364045d078364b39e4",
                "parameters": [
                    {
                        "name": "userId",
                        "in": "path",
                        "description": "ID do utilizador",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Lista de livros do utilizador",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "books_read": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "id": {
                                                        "type": "integer",
                                                        "example": 1
                                                    },
                                                    "title": {
                                                        "type": "string",
                                                        "example": "O Grande Livro"
                                                    },
                                                    "description": {
                                                        "type": "string",
                                                        "example": "Descrição do livro"
                                                    },
                                                    "read_time": {
                                                        "type": "integer",
                                                        "example": 15
                                                    },
                                                    "access_level": {
                                                        "type": "integer",
                                                        "example": 1
                                                    },
                                                    "is_active": {
                                                        "type": "boolean",
                                                        "example": true
                                                    },
                                                    "progress": {
                                                        "type": "integer",
                                                        "example": 75
                                                    },
                                                    "read_date": {
                                                        "type": "string",
                                                        "format": "date-time"
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        },
                                        "favorite_books": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "id": {
                                                        "type": "integer",
                                                        "example": 2
                                                    },
                                                    "title": {
                                                        "type": "string",
                                                        "example": "Outro Livro"
                                                    },
                                                    "description": {
                                                        "type": "string",
                                                        "example": "Descrição do livro"
                                                    },
                                                    "is_active": {
                                                        "type": "boolean",
                                                        "example": true
                                                    },
                                                    "favorited_at": {
                                                        "type": "string",
                                                        "format": "date-time"
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Utilizador não encontrado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Utilizador não encontrado"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
...
}